requirements.txt
streamlit
numpy
pandas
scipy
Pillow
nd2
matplotlib
